added formatting and view
