+++ cas-server-webapp-support/src/main/java/org/jasig/cas/web/report/HealthCheckController.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.jasig.cas.web;
+package org.jasig.cas.web.report;
 
 import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
