+++ cas-server-webapp-support/src/main/java/org/jasig/cas/web/report/StatisticsController.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.jasig.cas.web;
+package org.jasig.cas.web.report;
 
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.health.HealthCheckRegistry;
