+++ cas-server-webapp/src/main/webapp/WEB-INF/cas.properties
@@ -139,6 +139,12 @@ host.name=cas01.example.org
 # metrics.refresh.internal=30s
 
 ##
+# Reports
+#
+# Setting to whether include the ticket granting ticket id in the report
+# sso.sessions.include.tgt=false
+
+##
 # Password Policy
 #
 # Warn all users of expiration date regardless of warningDays value.
