+++ cas-server-webapp/src/main/webapp/WEB-INF/view/jsp/monitoring/viewStatistics.jsp
@@ -118,10 +118,11 @@
 </table>
 
 <h2>Performance Statistics</h2>
-
-<a href="statistics/ssosessions">SSO Sessions</a>
 <a href="statistics/metrics?pretty=true">Metrics</a>
 <a href="statistics/ping?pretty=true">Ping</a>
 <a href="statistics/threads?pretty=true">Thread Dump</a>
+
+<h2>Reports</h2>
+<a href="statistics/ssosessions">SSO Sessions</a>
 
 <%@include file="/WEB-INF/view/jsp/default/ui/includes/bottom.jsp" %>
