+++ cas-server-webapp/src/main/webapp/WEB-INF/web.xml
@@ -149,11 +149,6 @@
         <servlet-class>com.codahale.metrics.servlets.ThreadDumpServlet</servlet-class>
     </servlet>
 
-    <servlet>
-        <servlet-name>ssosessions</servlet-name>
-        <servlet-class>com.sun.jersey.spi.spring.container.servlet.SpringServlet</servlet-class>
-    </servlet>
-
     <servlet-mapping>
         <servlet-name>cas</servlet-name>
         <url-pattern>/login</url-pattern>
@@ -210,6 +205,11 @@
     </servlet-mapping>
 
     <servlet-mapping>
+        <servlet-name>cas</servlet-name>
+        <url-pattern>/statistics/ssosessions</url-pattern>
+    </servlet-mapping>
+
+    <servlet-mapping>
         <servlet-name>metrics-ping</servlet-name>
         <url-pattern>/statistics/ping</url-pattern>
     </servlet-mapping>
@@ -227,11 +227,6 @@
     <servlet-mapping>
         <servlet-name>metrics-health</servlet-name>
         <url-pattern>/statistics/healthcheck</url-pattern>
-    </servlet-mapping>
-
-    <servlet-mapping>
-        <servlet-name>ssosessions</servlet-name>
-        <url-pattern>/statistics/ssosessions</url-pattern>
     </servlet-mapping>
 
     <servlet-mapping>
